chore: Simplify auth feature
